M everything: seting files up to norms (done 3/4)
M player_move_translation.c: now moving X and Y independantly:
				won't block against walls anymore

+++ include/struct.h
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   struct.h                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ekrebs <ekrebs@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/05/28 09:37:14 by ekrebs            #+#    #+#             */
+/*   Updated: 2025/05/28 09:39:20 by ekrebs           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef STRUCT_H
 # define STRUCT_H
 
@@ -18,41 +30,41 @@ typedef struct s_parsing_map
 
 typedef struct s_ray
 {
-	double	dir_x;
-	double	dir_y;
-	double	side_dist_x;
-	double	side_dist_y;
-	double	delta_dist_x;
-	double	delta_dist_y;
-	int		map_x;
-	int		map_y;
-	int		step_x;
-	int		step_y;
-	int		hit;
-	int		side; // 0 = x, 1 = y
-	int		draw_end;
-	int		draw_start;
-	double realdist;
-	double perp_wall_dist;
-	int line_height;
-	mlx_texture_t *current_texture;
-	uint8_t *pixel;
-	uint32_t color;
-	double wall_hit_x;  // Exact X hit position
-    double wall_hit_y;
+	float			dir_x;
+	float			dir_y;
+	float			side_dist_x;
+	float			side_dist_y;
+	float			delta_dist_x;
+	float			delta_dist_y;
+	int				map_x;
+	int				map_y;
+	int				step_x;
+	int				step_y;
+	int				hit;
+	int				side; // 0 = x, 1 = y
+	int				draw_end;
+	int				draw_start;
+	float			realdist;
+	float			perp_wall_dist;
+	int				line_height;
+	mlx_texture_t	*current_texture;
+	uint8_t			*pixel;
+	uint32_t		color;
+	float			wall_hit_x; // Exact X hit position
+	float			wall_hit_y;
 }	t_ray;
 
-typedef	struct s_colors
+typedef struct s_colors
 {
-	char	*color_floor;
-	char	*color_ceiling;
-	int		floor;
-	int		ceiling;
+	char		*color_floor;
+	char		*color_ceiling;
+	int			floor;
+	int			ceiling;
 	uint32_t	color_floor_uint;
 	uint32_t	color_ceiling_uint;
 }	t_colors;
 
-typedef	struct s_path
+typedef struct s_path
 {
 	char	*north;
 	char	*south;
@@ -60,7 +72,7 @@ typedef	struct s_path
 	char	*west;
 }	t_path;
 
-typedef	struct s_map
+typedef struct s_map
 {
 	char	**grid;
 	int		x_max;
@@ -68,30 +80,29 @@ typedef	struct s_map
 	int		tile_size;
 }	t_map;
 
-typedef	struct s_image
+typedef struct s_image
 {
 	mlx_image_t			*img_window;
 	mlx_image_t			*player_img;
 	mlx_image_t			*minimap;
 }	t_image;
 
-typedef	struct s_texture
+typedef struct s_texture
 {
-	mlx_texture_t *north;
-    mlx_texture_t *south;
-    mlx_texture_t *east;
-    mlx_texture_t *west;
+	mlx_texture_t	*north;
+	mlx_texture_t	*south;
+	mlx_texture_t	*east;
+	mlx_texture_t	*west;
 	mlx_texture_t	*player;
 	t_image			*image;
 }	t_texture;
 
-
 typedef struct s_player
 {
-	double	position[DIM];
+	float	position[DIM];
 	char	facing;
-	double	direction[DIM];
-	double	radian;
+	float	direction[DIM];
+	float	radian;
 }	t_player;
 
 typedef struct s_toggles
@@ -119,6 +130,6 @@ typedef struct s_data
 	t_game			*game;
 	t_colors		*colors;
 	t_path			*path;
-} t_data;
+}	t_data;
 
 #endif
